requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
pymongo==4.3.3